5.31  测试查询(系统和基本信息)

+++ management/src/main/java/com/dataset/management/rest/DataSetController.java
@@ -5,6 +5,7 @@ import com.dataset.management.common.ApiResult;
 import com.dataset.management.common.ResultUtil;
 import com.dataset.management.consts.DataSetConsts;
 import com.dataset.management.entity.DataSet;
+import com.dataset.management.entity.DataSetFile;
 import com.dataset.management.entity.DataSystem;
 import com.dataset.management.service.*;
 import org.slf4j.Logger;
@@ -12,7 +13,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
+import scala.tools.cmd.gen.AnyVals;
 
 import java.io.IOException;
 import java.text.SimpleDateFormat;
@@ -52,7 +55,7 @@ public class DataSetController {
 
     //查询  Id
     @ResponseBody
-    @RequestMapping(value = "/listInfo/{dataSetId}",method = RequestMethod.GET)
+    @RequestMapping(value = "/selectById/{dataSetId}",method = RequestMethod.GET)
     public ApiResult listInfoDataSetByDataSetId(@PathVariable("dataSetId") int dataSetId){
         logger.info("开始罗列数据据基本信息");
         DataSet dataSet = dataSetService.findById(dataSetId);
@@ -64,8 +67,8 @@ public class DataSetController {
 
     //查询  datasetName
     @ResponseBody
-    @RequestMapping(value = "/listInfo/{dataSetEnglishName}",method = RequestMethod.GET)
-    public ApiResult listInfoDataSetByDataSetName(@PathVariable("dataSetEnglishName") String dataSetEnglishName){
+    @RequestMapping(value = "/selectByDataSetEnglishName/{dataSetEnglishName}",method = RequestMethod.GET)
+    public ApiResult listInfoDataSetByDataSetEnglishName(@PathVariable("dataSetEnglishName") String dataSetEnglishName){
         logger.info("开始罗列数据据基本信息");
         DataSet dataSet = dataSetService.findByDataSetEnglishName(dataSetEnglishName);
         if(dataSet.getId() ==0 ){
@@ -76,7 +79,7 @@ public class DataSetController {
 
     //查询  userName
     @ResponseBody
-    @RequestMapping(value = "/listInfo/{UserName}",method = RequestMethod.GET)
+    @RequestMapping(value = "/selectByUserName/{UserName}",method = RequestMethod.GET)
     public ApiResult listInfoDataSetByUserName(@PathVariable("UserName") String userName){
         logger.info("开始依据用户名【 "+userName+" 】罗列数据据基本信息");
         List<DataSet> dataSets = dataSetService.findByUserName(userName);
@@ -86,7 +89,7 @@ public class DataSetController {
 
     //查询全部
     @ResponseBody
-    @RequestMapping(value = {"/listInfoAll/{dataSetSortBy},{dataSetSortType}"},method = RequestMethod.GET)
+    @RequestMapping(value = "/selectAll/{dataSetSortBy},{dataSetSortType}",method = RequestMethod.GET)
     public ApiResult selectAllDataSet(@PathVariable(value = "dataSetSortBy") String sortBy,
                                       @PathVariable(value = "dataSetSortType") String sortType) throws IOException{
         Sort sort;
@@ -106,8 +109,15 @@ public class DataSetController {
             sort = basicSortBy();
             logger.info("按照默认方式排序");
         }
-        logger.info("开始罗列所有数据集系统表：");
+
         List<DataSet> dataSets = dataSetService.findAll(sort);
+        logger.info("更改所有数据集排序规则：");
+        for(DataSet dataSet: dataSets){
+            dataSet.setDataSetSortType(sortType);
+            dataSet.setDataSetSortBY(sortBy);
+            dataSetService.save(dataSet);
+        }
+        logger.info("开始罗列所有数据集系统表：");
         return ResultUtil.success(dataSets);
     }
 
@@ -141,18 +151,30 @@ public class DataSetController {
         int id = dataSet.getId();
         logger.info("获取修改的数据集ID："+id);
         /***
-         * {"id":1,"dataSetEnglishName":"sss","dataSetStatus":"COMPLETE",
-         * "dataSetPower":"PRIVATE","dataSetLastUpdateTime":"rrr",
-         * "dataSetUpdateDesc":null,"dataSetSortBY":null,"dataSetSortType":null,
-         * "dataSetSize":150,"dataSetBasicDesc":null,"dataSetCreateTime":null,
-         * "dataSetName":"sss","dataSetStoreUrl":null,"dataSetHiveTableName":null,
-         * "dataSetHiveTableId":null,"datatype":null,"dataSetFileCount":0
-         *
+         * {"id":23,"dataSetEnglishName":"bbbbb","dataSetStatus":"COMPLETE","dataSetPower":"PRIVATE","dataSetLastUpdateTime":"rrr","dataSetUpdateDesc":null,"dataSetSortBY":null,"dataSetSortType":null,"dataSetSize":150,"dataSetBasicDesc":null,"dataSetCreateTime":null,"dataSetName":"sss","dataSetStoreUrl":null,"dataSetHiveTableName":null,"dataSetHiveTableId":null,"datatype":null,"dataSetFileCount":0}
+
          */
-        dataSetService.save(dataSet);
-        DataSystem dataSystem = packageDataSystem(dataSet);
-        dataSetOptService.save(dataSystem);
-        return null;
+        long timetmp = System.currentTimeMillis();
+        SimpleDateFormat sdf=new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+        String newTime = sdf.format(new Date(Long.parseLong(String.valueOf(timetmp))));
+        dataSet.setDataSetLastUpdateTime(newTime);
+
+        String newDesc = "update the dataset "+dataSet.getDataSetName();
+        dataSet.setDataSetUpdateDesc(newDesc);
+
+        dataSet.setDataSetFileCount(0);
+        DataSet newDataSet = dataSetService.save(dataSet);
+        logger.info("更新后的数据集基本表："+newDataSet);
+
+        int oldDatasetId = dataSet.getId();
+        DataSystem old = dataSetOptService.findByDataSetId(oldDatasetId);
+        int oldId = old.getId();
+        logger.info("删除原有数据集系统关联表："+oldId);
+        dataSetOptService.deleteById(oldId);
+        DataSystem newDataSystem = packageDataSystem(dataSet);
+        logger.info("重新生成关联数据集系统表：");
+        dataSetOptService.save(newDataSystem);
+        return ResultUtil.success(newDataSet);
     }
 
     /**
@@ -160,11 +182,16 @@ public class DataSetController {
      * 删除了文件  files   更新 数据集基本表中关于文件的统计   filesCounts
      * */
     @ResponseBody
+    @Transactional
     @RequestMapping(value = "/clean/{dataSetId}",method = RequestMethod.POST)
     public ApiResult cleanDataSet(@PathVariable(value = "dataSetId") int dataSetId)throws IOException{
         DataSet dataSet = dataSetService.findById(dataSetId);
         logger.info("获取数据集名称："+dataSet.getDataSetName());
         logger.info("数据集准备清空操作  （删除文件） ");
+        List<DataSetFile> fileList = dataSetFileService.findDataSetFilesByDataSetId(dataSetId);
+        if(fileList.size() == 0 ){
+            return ResultUtil.error(-1,"此数据集不存在文件");
+        }
         dataSetFileService.deleteDataSetFilesByDataSetId(dataSetId);
         if(dataSetFileService.count() == 0){
             logger.info("数据集当前文件数："+dataSetFileService.count());
@@ -180,14 +207,12 @@ public class DataSetController {
             String newTime = sdf.format(new Date(Long.parseLong(String.valueOf(timetmp))));
             dataSet.setDataSetLastUpdateTime(newTime);
 
-            String newDesc = "清空了数据集所有文件。。";
+            String newDesc = "clean the dataset: "+ dataSet.getDataSetName();
             dataSet.setDataSetUpdateDesc(newDesc);
 
             dataSet.setDataSetFileCount(0);
             dataSetService.save(dataSet);
-            logger.info("数据及基本表新信息："+dataSet);
-            DataSystem dataSystem = packageDataSystem(dataSet);
-            dataSetOptService.save(dataSystem);
+            logger.info("更新后的数据及基本表新信息："+dataSet);
             return ResultUtil.success();
         }
         return ResultUtil.error(-1,"清空失败");
@@ -199,7 +224,7 @@ public class DataSetController {
     }
 
     private Sort basicSortBy(){
-        return changSortBy(DataSetConsts.SORTTYPE_ASC,DataSetConsts.SORT_BY_DATASET_ENGLISH_NAME);
+        return new Sort(Sort.Direction.fromString(DataSetConsts.SORTTYPE_ASC),DataSetConsts.SORT_BY_DATASET_ENGLISH_NAME);
     }
 
     private DataSystem packageDataSystem(DataSet dataSet){
